requirements.txt
alembic==1.10.3
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
filelock==3.10.7
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
numpy==1.24.2
pandas==2.0.0
protobuf==3.20.3
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
requests-file==1.5.1
six==1.16.0
SQLAlchemy==2.0.8
tldextract==3.4.0
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
urllib3==1.26.15
Werkzeug==2.2.3
zope.event==4.6
zope.interface==6.0
